FB-28-feat weekInfo Id를 이용한 해당 주차 통계 상세 정보 API

+++ src/main/java/mono/focusider/domain/article/dto/info/ReadingStatDetailInfo.java
@@ -5,6 +5,6 @@ import java.time.LocalDateTime;
 public record ReadingStatDetailInfo(
         LocalDateTime readingDate,
         Long readingTime,
-        Long understanding
+        Integer understanding
 ) {
 }
